Split schema from 1 index file into 2 separate for userSchema and noteSchema
