+++ server/src/models/user.ts
@@ -1,11 +1,11 @@
 import { InferSchemaType, model, Schema } from "mongoose";
 
-const userSchema = new Schema({
+const userModel = new Schema({
   username: { type: String, required: true, unique: true },
   email: { type: String, required: true, unique: true, select: false },
   password: { type: String, required: true, select: false },
 });
 
-type User = InferSchemaType<typeof userSchema>;
+type User = InferSchemaType<typeof userModel>;
 
-export default model<User>("User", userSchema);
+export default model<User>("User", userModel);
